rename temporary geo2.js -> geo.js :tada:
